authorization

+++ Angular/diff-test1/backend/routes/posts.js
@@ -43,7 +43,8 @@ multer({storage: storage}).single("image"), (req, res, next) => {
   const post = new Post({
     title: req.body.title,
     content: req.body.content,
-    imagePath: url + "/images/" + req.file.filename
+    imagePath: url + "/images/" + req.file.filename,
+    creator: req.userData.userId
   });
 
   post.save().then(createdPost => {
@@ -64,7 +65,7 @@ multer({storage: storage}).single("image"), (req, res, next) => {
 });
 
 // put new resource to override, or patch (existing)
-router.put("/:id", multer({storage: storage}).single("image"), (req, res, next) => {
+router.put("/:id", checkAuth, multer({storage: storage}).single("image"), (req, res, next) => {
   // console.log(req.file);
   let imagePath = req.body.imagePath;
   if (req.file){
@@ -75,12 +76,18 @@ router.put("/:id", multer({storage: storage}).single("image"), (req, res, next) 
     _id: req.body.id,
     title: req.body.title,
     content: req.body.content,
-    imagePath: imagePath
+    imagePath: imagePath,
+    creator: req.userData.userId
   });
   // takes js object and
-  Post.updateOne({_id: req.params.id}, post).then(result => {
+  Post.updateOne({_id: req.params.id, creator: req.userData.userId}, post).then(result => {
     // console.log(result);
-    res.status(200).json({message: "Update successful!"});
+    if(result.nModified > 0){
+      res.status(200).json({message: "Update successful!"});
+    }
+    else{
+      res.status(401).json({message: "Not authorized!"});
+    }
   });
 });
 
@@ -139,11 +146,14 @@ router.get("/:id", (req, res, next) => {
 router.delete("/:id", (req, res, next) => {
   console.log(req.params.id);
 
-  Post.deleteOne({_id: req.params.id}).then(result => {
-    console.log(result);
-    res.status(200).json(
-      { message: "Post deleted!"}
-    );
+  Post.deleteOne({_id: req.params.id, creator: req.userData.userId}).then(result => {
+    // console.log(result);
+    if(result.n > 0){
+      res.status(200).json({message: "Deletion successful!"});
+    }
+    else{
+      res.status(401).json({message: "Not authorized!"});
+    }
   });
